Model_Deployment/quantized_api_codebase/requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
huggingface_hub
python-dotenv
